Telecom fix and floating number addition or subtraction of wallet and payment due added

+++ src/components/client-popUp/clientPopUp.component.jsx
@@ -14,7 +14,7 @@ const ClientPopUp = ({clientName,purpose,id,updateStart,updateWalletStart,update
     const onAmountChange = (event)=>
     {
         let {value}=event.target
-        value= parseInt(value)
+        value= parseFloat(value)
         setClientId(id)
         setAmount(value)
     }
